moving effect size plotting to a new file; effect size config for dNLS cleaned - min samples in batch set to 200; multipletest correction was extracted from the plot code to the analyzer code
