+++ manuscript/run_all_manuscript_figures.sh
@@ -189,8 +189,8 @@ finetunedModel_MLPHead_acrossBatches_B56789_80pct_frozen=/home/projects/hornstei
 # -a $finetunedModel_MLPHead_acrossBatches_B56789_80pct_frozen $data_configs/AlyssaCoyneNEWEffectsFigureConfig $plot_configs/DistancesAlyssaCoynePlotConfig -q short -j alyssa_dist
 
 # ####### new dnls ###########
-# $NOVA_HOME/runnables/run.sh $NOVA_HOME/runnables/plot_effect_sizes -m 1000 -a \
-# $finetunedModel_MLPHead_acrossBatches_B56789_80pct_frozen $data_configs/dNLSNewEffectsFigureConfig  $plot_configs/DistancesdNLSPlotConfig -q short -j dnls_new
+$NOVA_HOME/runnables/run.sh $NOVA_HOME/runnables/plot_effect_sizes -m 1000 -a \
+$finetunedModel_MLPHead_acrossBatches_B56789_80pct_frozen $data_configs/dNLSNewEffectsFigureConfig_CLEAN  $plot_configs/DistancesdNLSPlotConfig -q short -j dnls_new
 
 ##### neurons ########
 # $NOVA_HOME/runnables/run.sh $NOVA_HOME/runnables/plot_effect_sizes -m 1000 -a \
@@ -202,5 +202,5 @@ finetunedModel_MLPHead_acrossBatches_B56789_80pct_frozen=/home/projects/hornstei
 
 
 
-$NOVA_HOME/runnables/run.sh $NOVA_HOME/runnables/generate_umaps_and_plot -m 5000 \
--a $finetunedModel_MLPHead_acrossBatches_B56789_80pct_frozen $data_configs/AlyssaCoyneNEWUMAP0FigureConfig $plot_configs/UMAP0AlyssaCoyneColorByGroupPlotConfig -q short -j umap0_alyssa
+# $NOVA_HOME/runnables/run.sh $NOVA_HOME/runnables/generate_umaps_and_plot -m 5000 \
+# -a $finetunedModel_MLPHead_acrossBatches_B56789_80pct_frozen $data_configs/AlyssaCoyneNEWUMAP0FigureConfig $plot_configs/UMAP0AlyssaCoyneColorByGroupPlotConfig -q short -j umap0_alyssa
