+++ src/analysis/analyzer_effects_alyssa_OLD.py
@@ -32,6 +32,8 @@ class AnalyzerEffectsAlyssaOLD(AnalyzerEffectsDistRatio):
         embeddings_df = self._prepare_embeddings_df(embeddings, labels)
         embeddings_dim = embeddings.shape[1]
         effects_df = self._calculate_all_effects(embeddings_df, baseline_perturb_dict, embeddings_dim, n_boot)
+        self._correct_for_multiple_hypothesis(effects_df)
+
         self.features = effects_df
         return effects_df
